fixed sh variable control issues

+++ Project3/ARR/SampleScene.hpp
@@ -20,7 +20,7 @@ enum DrawMode
 
 struct Block
 {
-	float N;
+	int N;
 	float* hammersley;
 };
